Work around bug in SDK 10.13 regarding UUID
See https://forums.developer.apple.com/thread/84375
In 10.12 and below, the `identifier` (UUID) property of a Bluetooth
device is in `CBPeripheral`. In 10.13, the `identifier` was moved into a
new `CBPeer` class, from which `CBPeripheral` now inherits.

Unfortunately, the `identifier` property of `CBPeer` is tagged as being
available only in 10.13 and above, which means that even though
`peripheral.identifier` is the correct way to access that property in
all SDK versions, it causes problems when building for 10.12 or below
while using a 10.13 SDK.

This workaround accesses the `identifier` through the peripheral
object's key-value interface, bypassing the deprecation warning. I'm
certainly open to suggestions regarding a better workaround.