Add WebSocket message defrag & decode in Session
The .NET `WebSocket` class returns message frames, and each such frame
may or may not represent the final frame in a message. The Session base
class now assembles fragments if necessary, decoding them from UTF-8 as
it goes. This is a little different from the Mac side, but I expect the
hardware-specific session API to be equivalent to that on the Mac.