Windows: Implement Session base class & example
`Session` is a base class for hardware interface-specific device session
types, just as it is in the Swift version of this application. The
primary difference is that the C# version uses `async` and `await`
internally instead of Swift's completion handler pattern. The API
between `Session` and derived classes, though, does use the completion
handler pattern since it allows the session precise control over the
ordering of a `Request`'s response relative to any other messages the
handler may send.

The `BLESession` in this commit is only here to serve as an example, and
in particular demonstrates the response message timing issue discussed
above.