Correctly serialize error response
The `JProperty(string, object)` constructor assumes that the `object`
parameter is already a JSON type (like `JObject`, `JValue`, etc.) -- in
particular, it doesn't call any conversion function on the object passed
in. This was leading to an exception ("Could not determine JSON object
type for type") when trying to report errors to the remote caller.
Calling `JObject.FromObject(object)` allows `JsonRpcExceptionConverter`
to get in on the action to correctly serialize the error object.