Makefile enhancements: CONFIG var, "dist" target
The "CONFIG" variable in the Makefile can be changed to "release" or
"debug" to compile a release/debug build.

The "all" target no longer generates a PKG file and instead just builds
the app. The new "dist" target builds a PKG with a filename consistent
with the version number stored in the app.