Use Fleck for WSS on Windows
Using .NET's `HTTPListener` for WSS is possible but requires jumping
through quite a few hoops, such as adding a certificate to the OS
certificate store. That action requires a UAC prompt and users in a
managed situation (schools, etc.) may not have rights to do so.

On the other hand, Fleck can run WSS using a certificate loaded in any
of a number of ways. I chose to load it through a bundled resource.

Conversion required some rework, of course, but it turns out I like the
Fleck API better anyway ;)