Ensure session is ready before handling messages
Apparently Fleck doesn't guarantee that `OnMessage` and `OnBinary` won't
be called until after `OnOpen` -- rather, it seems to be a race
condition. Even better, in Scratch Link's case the race seems to be
partially determined by whether or not the application is packaged for
the store.

This change creates the session object as soon as the WebSocket connects
instead of waiting until `OnOpen`, which guarantees that the session
object will exist before any methods try to use it.