Register ValueChanged before requesting Notify
Apparently, notifications can come in VERY soon after writing the
`Notify` flag to the characteristic descriptor, and on a different
thread. This was leading to missed notifications between making the
request and registering the event handler. Now we register the event
handler first, then unregister it if the notification request fails.

I also added a bit of locking in `Dispose()` to avoid a shutdown-time
exception.