make details label more descriptive
When the conflicting program is another instance of Scratch Link, just
using "Details" as the label could be confusing: is it showing details
about the current program or the conflicting program? This change will
hopefully make it clear that the "Details about the conflicting program"
are about the other instance.