move to port 20111; cleanup
Using the same port as older SSL Scratch Link makes it difficult to get
good fallback behavior in the browser: a not-yet-failed `wss://`
connection can block a `ws://` connection to the same port which would
have succeeded otherwise, and vice versa. This can even apply to
recently closed sockets, which complicates the user experience even
further. Leaving 20110 for `wss://` and using 20111 for `ws://`
eliminates these conflicts.