refactor EventAwaiter for better reliability
This new structure better supports code which wants to await an event
that it triggers. With the previous structure it was possible to miss
the event, since the event could fire after the code which triggers it
but before the event is hooked. This was especially problematic when
debugging.

Another bonus of this new structure is that it's possible to `await` the
same event multiple times, even in a loop, without rebuilding the
`EventAwaiter` and without any chance of missing events.