feat(Safari): inject client script into page if script ID is present
The presence of a script element with ID "scratch-link-extension-script"
is interpreted by the Safari extension as a request to inject its client
script into that element. The extension automatically looks for that
element when its content script runs (when the document is done loading)
but the page can also request it later by posting a message with a
truthy property called "inject-scratch-link-script".