fix(safari): don't cause Safari to steal focus for every Scratch Link -> page message
If the native app sends a message to the background script, Safari forces its window to the top and
steals focus. This change polls for messages instead, which is maybe not great but at least it's
less bad than Safari stealing focus. On the plus side, this takes the place of the keepalive
messages. Also, the native app portion of the extension now supports non-integer message IDs.