fix: fix minor MAS compliance issues
* Add `LSApplicationCategoryType` key to the app's `Info.plist`
* Fix capitalization of `PlugIns` directory for embedding the extension
  into the app
* Fix helper & extension bundle IDs to start with the app's bundle ID

Note that the helper now uses the same product bundle ID as the Scratch
Link app. This is necessary to successfully build the extension with the
helper. The "Scratch Link Safari Helper" can now be thought of as the
"Scratch Link Safari Extension build helper"... hopefully I can find a
more elegant solution in the future.